// src/api/openapi/models/homework-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API
 * Class APIs are served under /api/v5 while legacy domains may still remain on /api/v1 during migration.
 *
 * The version of the OpenAPI document: v5-class-migration
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EvaluationResponse } from './evaluation-response';
// May contain unused imports in some cases
// @ts-ignore
import type { HomeworkContentDto } from './homework-content-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { ImageDto } from './image-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { PeerReviewResponse } from './peer-review-response';

export interface HomeworkResponseDto {
    'homeworkId'?: number;
    'submitterId'?: number;
    'submitterNickname'?: string;
    'submitterProfileImage'?: ImageDto;
    'requestedByStudyLeader'?: boolean;
    'submitted'?: boolean;
    'submissionTime'?: string;
    'homeworkContent'?: HomeworkContentDto;
    'evaluation'?: EvaluationResponse;
    'peerReviews'?: Array<PeerReviewResponse>;
}