// src/api/openapi/models/grade-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API v1
 * v1
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface GradeDto {
    'gradeId'?: number;
    'code'?: string;
    'score'?: number;
    'label'?: string;
    'type'?: GradeDtoTypeEnum;
    'orderNum'?: number;
}

export const GradeDtoTypeEnum = {
    Mission: 'MISSION'
} as const;

export type GradeDtoTypeEnum = typeof GradeDtoTypeEnum[keyof typeof GradeDtoTypeEnum];